// types.ts
export enum Format {
  // Common Text & Markup Formats
  Markdown = 'Markdown',
  CommonMark = 'CommonMark',
  MarkdownStrict = 'Markdown (Strict)',
  MarkdownPHPExtra = 'Markdown (PHP Extra)',
  MarkdownGithub = 'GitHub-Flavored Markdown',
  MarkdownMMD = 'MultiMarkdown',
  PlainText = 'Plain Text',
  HTML = 'HTML',
  XHTML = 'XHTML',
  RichText = 'Rich Text Format (RTF)',
  LaTeX = 'LaTeX',
  ConTeXt = 'ConTeXt',
  Beamer = 'Beamer',

  // Wiki & Lightweight Markup
  RST = 'reStructuredText',
  OrgMode = 'Org Mode',
  AsciiDoc = 'AsciiDoc',
  MediaWiki = 'MediaWiki',
  Textile = 'Textile',
  Creole = 'Creole',
  Muse = 'Muse',
  Txt2Tags = 'Txt2Tags',
  JiraWiki = 'Jira Wiki',

  // XML-based Formats
  DocBook = 'DocBook',
  TEI = 'TEI',
  JATS = 'JATS',
  OPML = 'OPML',
  ICML = 'ICML',
  
  // Data & Serialization Formats
  JSON = 'JSON',
  YAML = 'YAML',
  CSV = 'CSV',
  TSV = 'TSV',
  BibTeX = 'BibTeX',
  BibLaTeX = 'BibLaTeX',
  CSLJSON = 'CSL JSON',
  CSLYAML = 'CSL YAML',

  // Binary Formats
  ePub = 'ePub',
  DOCX = 'Microsoft Word (docx)',
  ODT = 'OpenDocument Text (ODT)',
  PDF = 'PDF',
}